Add optional `managed` parameter to device c’tor
Sometimes we might want to create a `Device` instance without it being managed by the device managed. This is for example, if we simply want to get a few attributes of a device if all we have is the MAC address. In that case, we create a Device instance, read the attributes and then forget about the instance.

Particularly of interest for sub classes of `DeviceManager` who want to decide on certain device properties if they then create a subclass instance of that `Device` or not.